feat(api): Phase 8 — Polish & QA
- Remove all remaining localhost:3000 / json-server references
- Create centralized date utilities (timeAgo, formatDate, formatDateTime)
- Update NavBar to use shared timeAgo helper
- Verify clean build across all phases
