+++ src/components/NavBar.vue
@@ -268,16 +268,9 @@ async function handleLogout() {
     router.push('/');
 }
 
-// ── Helpers ───────────────────────────────────────────────────
-function timeAgo(dateStr) {
-    const diff = Date.now() - new Date(dateStr).getTime();
-    const mins = Math.floor(diff / 60000);
-    const hours = Math.floor(diff / 3600000);
-    const days = Math.floor(diff / 86400000);
-    if (mins < 60) return `${mins}m ago`;
-    if (hours < 24) return `${hours}h ago`;
-    return `${days}d ago`;
-}
+import { timeAgo } from '@/utils/date';
+
+// ── Helpers moved to @/utils/date ────────────────────────────
 
 // ── Click outside — closes both dropdowns ─────────────────────
 function handleClickOutside() {
